+++ src/main/java/com/g3/user/service/interfaces/IUserService.java
@@ -9,10 +9,10 @@ import com.g3.user.model.User;
 public interface IUserService {
     public User register(User user) throws CpfOrEmailInUseException;
     public List<User> getAll();
-    public User update(User newUserData, Integer userId) throws UserNotFoundException;
-    public void delete(Integer id) throws UserNotFoundException;
+    public User update(User newUserData, Long userId) throws UserNotFoundException;
+    public void delete(Long id) throws UserNotFoundException;
     public List<User> searchByName(String name);
-    public User searchById(Integer id) throws UserNotFoundException;
+    public User searchById(Long id) throws UserNotFoundException;
     public List<User> searchByCPF(String CPF);
     public List<User> searchByEmail(String email);
 }
